\name{parseConfigFile}
\alias{parseConfigFile}
\title{Parse the config file of the DataCube}
\description{Reads the cube.config file from the datacube's base directory and returns a list with its content}
\usage{parseConfigFile(cubepath)}
\arguments{
  \item{cubepath}{Path to the datacube}
}

\value{A list with the cube config entries}

\author{Fabian Gans, Miguel D. Mahecha, MPI BGC Jena, Germany, fgans@bgc-jena.mpg.de mmahecha@bgc-jena.mpg.de
Fabian Gans}





